requirements.txt
django
djangorestframework
redis
pyyaml
uritemplate
tldextract

pytest
pytest-django
pytest-redis
mypy
